move SimpleCache.cs into Utils Folder
